Compute MAP
- Default settings from sbi work fine

+++ model_fitting/sbi/inference/potentials/posterior_based_potential.py
@@ -98,7 +98,7 @@ class PosteriorBasedPotential(BasePotential):
 
         with torch.set_grad_enabled(track_gradients):
             posterior_log_prob = self.posterior_estimator.log_prob(
-                theta, context=x_repeated
+                theta.float(), context=x_repeated
             )
 
             # Force probability to be zero outside prior support.
